requirement.txt
torch>=1.2.0
torchvision==0.4.0
tensorflow==1.14.0
tensorboard==1.14.0
pillow